pyproject.toml
[project]
name = "query-cost-guard"
version = "0.1.0"
description = "Per-query cost guard for cloud databases. Set dollar or byte limits before any query runs."
license = {text = "MIT"}
requires-python = ">=3.11"
readme = "README.md"
keywords = ["bigquery", "cost", "guard", "query", "cloud"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
]
dependencies = [
    "google-cloud-bigquery~=3.27.0",
    "google-cloud-billing~=1.15.0",
    "pydantic~=2.10.0",
    "cachetools~=7.0.1",
    "capacity~=1.3.14",
    "structlog~=25.5.0",
]

[project.scripts]
query-cost-guard = "query_cost_guard.cli:app"

[project.optional-dependencies]
cli = ["typer~=0.24.0"]
dev = ["pytest", "pytest-mock", "ruff>=0.1.0", "pylint>=3.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "C4",
    "SIM",
]
ignore = [
    "E501",
]
